testsuite/test.sh: Re-enable running tests from src dir.
The existing code exits if it can't find host.conf in the working
directory.  This patch removes that exit, as this was preventing running
the test suite from the source directory since host.conf will never be in
the source directory.  There are overrides for variables essential to the
test process.