testsuite/test.sh: Add fetch tests section.
When run in the src dir the find_snapshot test would generate errors
because the md5sums file that find_snapshot uses wasn't downloaded.  This
test addition creates a temporary build directory and performs a "fetch
md5sums" test.

This patch also adds a wget quiet option that is used when running the
test suite.