lib/common.sh: Invoke find_snapshot as a subshell in get_source.
The current use of find_snapshot in get_source is erroneously directly
invoked and its output isn't properly trapped and pollutes the stdout of
get_source, which can make get_source erroneously return output in an
error case.

Putting this in place regresses the tar.bz2 test case.  This will be fixed
in a following commit.