lib/common.sh: find_snapshot should return list of found snapshots even on error.
Added a new test which tries to find an ambiguous snapshot.  This exposed a bug
in find_snapshot where it wasn't echoing the snapshots to stdout when there was
an error (such as too many matching snapshots).  This patches adds the new test
and fixes this error case.

This test also causes a return of '1' if too many matching snapshots were found
but one wasn't chosen, whereas in the past it would erroneously return '0'.