lib/make.sh: Replace hard coded dynamic linker version number.
When qemu-aarch64 is used for make check we need to replace the
dynamic-linker symlink in the sysroot lib directory with a copy of the
actual dynamic linker, but with the correct version number that matches the
clibrary version.  This patch replaces the dynamic linker version number
that is hard coded with a programmatically determined dynamic-linker
version number.

Change-Id: Iab15c7c36f10e9fea491af36c3939f96a008fab2