Add quotes around function inputs to properly expand built-in variables during assignment.
Depending on the shell version, assignments of built-in variables
without double quotes might generate an error.  For instance:

function_bar ()
{
    local foo;
    foo=$1
}

The expansion of $1 and assignment to foo might generate an error if
function_bar is called with a parameter value that contains spaces
(delimiters), e.g., function_bar "input one".  Adding double quotes
around the built-in variables will allow the assignment to include a
string.

Change-Id: I5b6fcfd0b254ef21d748f8c53a9885c0a88e0ee3