Check check_stamp for error code 255 and return failure in such cases.
The return code 255 from check_stamp indicates that the compare file is not
present, but it always should be since it's what the stamp is compared
against.

There is currently a FIXME that shows up in the build() stage the first
time build is called on a package, because we're checking against the
package srcdir, but the srcdir does not yet exist (unless the srcdir was
checked out before hand using --checkout all|<package>).

The existing code doesn't prevent the checkout from happening because we're
not failing on error code 255 (but we should be).

The second time build() is called on a package (a rebuild for instance) the
build stamp is compared against the package srcdir timestamp.  Unless the
user has manually updated the srcdir the buildstamp will always be newer
than the srcdir and the build won't be rerun.

This is a bug.  The srcdir should always be updated in build() before the
build stamp is compared against the srcdir timestamp.  In this way, the
build stamp timestamp will be older than the srcdir timestamp.

This patch corrects the following issues:

1) FIXME about missing compare_file the first time build() is called for
a package.

2) check_stamp returning 255 now results in a failure wherever it is used.

3) The build is rerun if the srcdir is newer than the build stamp.

Change-Id: I5988f6f744a2f68a2f10dd1c03eb2ab6d6f87b89