Add support for schroot testing
Functions for schroot-based testing are in new file lib/schroot.sh.
Schroot testing is a feature specific to a particular DejaGnu board,
if board loads schroot-ssh.exp file, then it is deemed to support
schroot-based testing and has been configured for it.
At the moment this is a no-op, as no board files include schroot-ssh.exp.

Change-Id: Id6145d83c64582a6d195eb113d86fba56992b365