Support git-over-ssh
Already submitted via bugzilla, but some later patches depend on
it, so included here for convenience.

Think I found all the right places to make updates in the source.

Dealt with testing mainly by looking for tests on git:// urls and
creating ssh:// versions of those tests. In many cases, did this
by introducing for loops that iterate over ssh, git, http for these
tests. Not sure that I've covered all the cases.

Updates to push() and commit() are untested as these functions
seem never to be called.

Change-Id: Ib216117f689615b4b7db8877f9cc54c4533c4021