Put make flags from config files at the end
This allows flags in the config file to override flags coming from
other parts of the script. This is useful if, for instance, there
is a race condition in the tool's build and we therefore need
parallelism always-off for that build.

An alternative approach would be to make all CLI-specifiable
features per-tool overridable by a special flag in the .conf,
e.g. parallel=no. This might be cleaner, but its a bit less
flexible and a bit more work to implement.

Added tests for make_all and make_install.

Change-Id: I6de20cbf46ad4913c5465872599de880d7d264e0