Get default_makeflags by source rather than grep
We add read_config to source the config file in a subshell and get
the variable we're interested in. Like the 'grep' approach, this
does not pollute the environment. Lile the 'source' approach, it
lets us write shell in the config file to alter variables depending
on conditions. The implementation is a bit complicted, which is
concerning, but it seems to work.

Change-Id: I31d12725c7e11e04506115a2595c0792638eb287