wip
bash scripts/benchmark.sh eembc panda works from start to end

Note that it does not do freq control (not supported, and dangerous, on panda)
It also requires a wrapper makefile around the benchmark. That's OK for
GNU toolchain makefile emulation, and mostly OK for 'invoke the benchmark
as we want it invoked' - except that we might not want to share that last
bit were these parts ever to go public. But it was the least fiddly way to do
it.

Next week:
(Monday: try to finish the lllh commentary first?)
Tidy up
Make it work on LAVA (panda)
Make it work on LAVA (juno) - may involve building an oe filesystem
Test freq ctl
Get eembc sources into access-controlled git