Ensure each net if is usable before returning from start_network
Or at least attempt to do so via ping. This seems like good practice,
and it fixes a bug where benchmark.sh would often fail to get anything
back while listening for the post-benchmarking-run IP address. Though
inserting equivalent code in benchmark.sh does not fix that problem,
so I'm not convinced I've got to the real bottom of this.

Change-Id: I54076aa240621eade668bdd1fccc0284ebea24b2