Use a script rather than a makefile to run the benchmark
The makefile rule is effectively an embedded shell script anyway,
and this way we avoid any problems around nesting make. Nesting
make is a significant problem in SPEC, for instance.

The only real problem is that this can lead to defining the same
variable in more than one place. TODO Eliminate this problem. This
probably just means making ITERATIONS static in the config file.
Note that I was already thinking about using substitutions more in
the config file.