Pass parameters to linarobench.sh
We pass bench_boardargs (static target info, embedded in
config/boards/bench/* files), then a --, then run_benchargs (args
given by the user when he invokes scripts/benchmark.sh).

linarobench.sh can be different for every benchmark, but the
expectation is that it will do something sensible with the target
info on the LFS of the -- (in the spec case, inserting it into the
config) and pass the arguments on the RHS through to whatever
linarobench is invoking (likely to be make for eembc, runspec for
spec).