Ignore ssh config when using LAVA
It seems that ControlMaster=yes causes horrible problems.

Not fully understood yet, but committing this in the hope that
it will unblock benchmarking runs.

I've not applied this change for non-LAVA uses in runbenchmark.sh,
but I _have_ applied it all over functions that might be used for
LAVA or non-LAVA processes.