Better parsing of srcdir into repodir
This replaces a line of code that worked out what to call
the repo directory (i.e. local place to put the master
branch) based on the src directory. That code just
chopped off branch and revision identifiers in a
simplistic way.

This code instead modifies srcdir into a local ssh git
URL and then uses our git parsing code to extract the desired
directory name.

The result is still imperfect - we have a problem if the path
contains '@' characters that do not refer to revision. Jenkins likes
to create such paths. The proper place to fix that is likely the
git parser.

Change-Id: I33aaec38b20009b8f5298606e62757e6f3fa65c1