Always invoke make_check() with 'binutils' or 'gdb' in $2 when checking binutils and gdb.
This fixes a bug introduced by d80db8525f8af5f385e897dca3d717ec98ecbf8f
which adds '-binutils' and '-gdb' to binutils and gdb builddir paths since
they use a merged git tree.  The problem is that not all functions that use
get_builddir() were calling it with a second parameter of 'binutils' or
'gdb' which is how ABE determines whether to append '-binutils' or '-gdb'.

This patch makes sure to call make_check() with 'binutils' or 'gdb' in $2
so that get_builddir() is always called with something in $2.

Change-Id: I9f91418480ab8597949439e782ae87f8ef639a0e