lib/common.sh: dryrun() should propogate the non-tee return value of 'eval $1'
The dryrun function, if called with an external command (that is likely to
fail) and piped into tee will have the return value of 'tee' returned from
dryrun.  The tee program will always return '0'.

It is not the intention of dryrun that the eval'ed command always return
success, but rather the result of the non-tee part of the expression,
e.g.,

   dryrun "ls <non-existant file> | tee _out"

This will return '0' when it should return 'non-zero'.

The solution is to wrap the 'eval $1' in dryrun in pipefail as follows:

(
    set -o pipefail
    eval $1
)
return $?

This will cause $? to hold the right-most, non-zero return value from the
eval'ed expression, otherwise it will hold '0'.

This technique should be used anywhere | tee is used that will squelch an
important return value.

Change-Id: I79c55e02369b083495e3aa256516e261169bfa68

Conflicts:

	testsuite/test.sh