Add helper scripts to compare the results of 2 Jenkins jobs.
CompareJobs.job is executed by the Jenkins server, while compare_jobs.sh is executed on the fileserver, where the logs are stored.

The results are compared target per target, so only the targets in common between the 2 builds are compared.

At present, it can only compare jobs which store their results with a similar layout on the server, such that only jobs of the same 'family' can be compared. This can probably be improved in the future, in case it is needed.

These scripts generate XML reports, displayed by Jenkins via the "Summary Display Plugin", for instance in the CompareJobs job.

Change-Id: Iffe4bb9797c07b9edbb695846cccb298f60b9109