ABE test.sh and testsuite/test.sh should return non-zero if there are test failures.
In order for test results to be useful in automation the ABE API tests
(test.sh) and unit-tests (testsuite/test.sh) should return non-zero if
there are any test-suite failures.  We have to be careful and not return
the number of failures only as this could overflow the return value to
zero.

Change-Id: I102f24891f8db45243d10c41e5727a7a239355ef