git parser should return proper 'tool' on tarballs with more than one dash in the name
Building a release candidate binary toolchain from the following tarball
name caused the parser to return the wrong tool name:

gcc-release-candidate-2015.05-rc2

It was returning 'gcc-release-candidate'.  The fix makes sure that only
'gcc' is returned, and adds a special case for 'binutils-gdb' if it shows
up in archive name.

Added two new tests to verify both cases, and existing cases in the parser
pass.

Change-Id: I5f838fc12046755b0111e39a1372f3129094c871