Switch aarch64-linux-gnu testing to use real HW
We can't use real hardware for all of aarch64-linux-gnu, arm-linux-gnueabi, and
arm-linux-gnueabihf testing because the board crashes.  We can't use QEMU
for aarch64-linux-gnu testing because QEMU gets out of control and fills the
system with its processes.  The current attempt leaves testing of armv7/aarch32
on QEMU (which works OK), and will use ARMv8 board for aarch64-linux-gnu
testing in hopes that it can deal with a third of the load.

Change-Id: I24dd35a991e5d975d88f55a23782d400f77e64b5