Remove last listener
Remove a bunch of support code that was only there for the listener,
too.

We replace this with code that sends a ping when the benchmark run
finishes on a target. On receipt of ping, the host will try to get
the benchmark data from the target that sent the ping (it assumes
that the hostname (or IP address, for nameless networks) is stable).
For protection against connecting to the wrong target, we require
that the public key for each target is in known_hosts before benchmarking
begins, and we set StrictHostKeyAuthentication to yes.

On the plus side, this is less complicated and thus less error-prone
and more maintainable. It's also more secure, in that our only service
is now ssh - not ssh + netcat.

On the minus side, it is less capable. We confine the scripts to
only working:
1) Within a single network (all machines must be able to ping one another)
2) With reliable addresses (stable IP or hostname)

If these conditions are not met, failure is reasonably visible: the
host will either not notice that the target completed its benchmark
(but the user might), or the host will get a message that a target
completed the benchmark from a wrong IP address and thus fail to
connect to it. Either way, the target will stay up, and so data can
be collected 'by hand', provided the necessary private key is availble.

Change-Id: Ieb9393a2b9d665eea268dab1ab24232e213360cc