lib/configure.sh: Set libdir=/usr/lib, slibdir=/lib, and rtlddir=/lib
Fix for https://bugs.linaro.org/show_bug.cgi?id=1717

ABE is to be consistent with Ubuntu library layouts because there is
historic precedent for this layout in Linaro binary toolchains.

"slibdir" stands for shared-objects libdir.  Therefore all shared
objects should go into slibdir.  On Ubuntu systems this is /lib.
"libdir" is where static libraries and linkers scripts are to be
installed.  On Ubuntu systems this is /usr/lib.  "rtlddir" is where
the dynamic-linker is to be installed.  On Ubuntu systems this is
/lib.

Change-Id: I19242cb2ce118505a7f53224e1956b738ad2cd85