lib/checkout.sh: prune git branches before git pull in checkout() 'Update' code path.
This is required due to the following scenario:  A git reference dir
is populated with a git clone on day X.  On day Y a developer removes
a branch and then replaces the same branch with a new branch of the
same name.  On day Z ABE is executed against the reference dir copy
and the git pull fails due to error: 'refs/remotes/origin/<branch>'
exists; cannot create 'refs/remotes/origin/<branch>'.  You have to
remove the stale branches before pulling the new ones.

Change-Id: I436b4961deb5496e3b28567bad59e22f91ee7d09