Fix bug #1845.
To distinguish tests with the same name in different components,
prefix the test names with the component name.

This happens for instance with tests common to gcc and g++. Since
we concatenate all the .sum files before comparison, the
compare_dg_tests.pl script could be confused by multiple occurrences
of the same test.

With this patch, such tests will be prefixed by 'gcc:' or 'g++:', eg:
gcc:c-c++-common/asan/swapcontext-test-1.c

Change-Id: I846a3f96dd9f272394e9814b5be53964ffb462db