Support uinstance jobs
Add a concept of 'trusted' or 'untrusted'. Untrusted jobs
work as before, but the logic is more clear now as it's all
conditionalized, and comments have been improved.

Trusted jobs assume that the host knows what it's doing, and
that we are happy to upload results to LAVA bundle stream.

Change-Id: Ia578040fd264e2175de961ac41f3f7cc4e4654d4