Clean up toolchain usage
Toolchain is always required - even with a prebuilt benchmark, we
have to know the triple to know which build to select from the
prebuit tarball. And we specify the triple in the form of the
toolchain identifier. However, if prebuilt is given the toolchain
does not have to actually exist, or be downloaded.

Be more permissive when calculating the triple, so that we can
enter strings like aarch64-linux-gnu-clang.

Tighten up/fix error checking in consequence.

Change-Id: I1a8627877aafbb3673af4db458f241019de7c53c