Decouple triple from toolchain
Allow user to specify triple independently of toolchain.
This lets us:
1) Not give a toolchain at all for prebuilt benchmark tarballs
2) Identify a toolchain to use in tarballs with more than one gcc

The triple is optional. Conditions:
Triple given, toolchain is binary: We use the toolchain pointed
              at, benchmark scripts understand it to be targeting
              the given triple.
Triple given, toolchain is tarball: We try to use ${triple}-gcc.
              If this does not exist, we try to use gcc. If gcc
              exists, benchmark scripts use it and understand it
              to be targeting the given triple. If gcc does not
              exist, we fail. In all cases, we are talking about
              a GCC inside the tarball, not from anywhere else.
Triple given, toolchain not given: We use a prebuilt tarball,
              understood to be targeting a system of the type
              described by triple.
Triple not given, toolchain is binary: We assume that the
              toolchain is native for the host.
Triple not given, toolchain is tarball: We use the binary in the
              tarball named gcc, which is expected to be native
              for the host.
              If the tarball does not contain a gcc binary, we
              fail.
Triple not given, toolchain not given: We use a prebuilt tarball,
              understood to be targeting native for the host.

Change-Id: Ie5d783299448b433c759aa6ec5235a4a44c69c54