Rationalize *-session YAML
host-session-multilib.yaml maps to the old host-session.yaml and is
suitable for use on KVM targets, where abe requires the relevant
packages to be able to configure.

host-session-no-multilib.yaml doesn't install the multilib packages
and is suitable for use on Juno (and other?) targets, where the
packages cannot be installed (and abe configure then works anyway,
for whatever reason).

host-session.yaml does not try to install any packages at all. This
is where we are trying to get to.

target-session-tools.yaml maps to the old target-session.yaml and
is used by targets that can install packages to make sure that they
have all of the tools that we like to use for benchmarking.

target-session.yaml does not try to install any packages at all.
This is where we are trying to get to and it is what we have to do
now for Mustang, where we are using OpenEmbedded.

Target selection logic updated in Benchmark.sh. We now default to
target-session-tools.yaml, but override for Mustang.

Change-Id: Iff978d1ab685a885a1419b2834490852908ecc55