Apply NAME=VALUE CLI overrides before Benchmark.sh
Previously these overrides were applied after Benchmark.sh, which
is handy for letting the user override anything at all, but is an
awkward behavioural deviation from the Jenkins use of Benchmark.sh,
where nothing gets overridden after the fact.

Make Benchmark.sh always the final arbiter of what goes in. Nearly
every value it sets will anyway respect any pre-set override.

This also makes it possible for Benchmark.sh to record all its
settings as metadata, which is the immediate motivation for making
this change. We really want to do this in Benchmark.sh so that
metadata is recorded in the same place, regardless of whether we
are using CLI or Jenkins interface.

Change-Id: I2f34763cf7002c4f1a96be212d5000f7a5b87a0d