Handle different spellings of benchmark names
The names are of form number.name, but when they are used in
identifiers in the raw file, they are in form number_name. Of
course, this worked by coincidence, because '.' matches any char
in a regexp.

Patch cleans this up by providing specialised lookup function for
names, which converts the name from 'dot' to 'underscore' form.

Change-Id: I95edb140561ae61a2aece6ae26f2bfe9a1502d6c