Handle possibility that gather script is a symlink
Our upload method preserves symlinks, so when we copy, say,
CPU2006.sh, we end up with a gather.sh that is a symlink to a file
that does not exist.

The ideal solution to this would be to make CPU200x.sh hardlinked
to CPU200{0,6}.sh, but git does not understand hard links.

We want to keep the core scripts ignorant of LAVA, so we cannot put
an entry in the config file for the benchmark.

But we can update host-session to be aware that the gather script
might be a symlink. Not nice, but the best I could come up with
that preserves the layering.

Change-Id: Ie6bb0ba035e09fa6e6064469ffc6f8117930fde1