Set 'build' in Coremark-Pro default_makeflags
All this time abe.sh has been using EEMBC.conf to build
CoremarkPro, because we had CoremarkPro on a branch of the EEMBC
repository.

It turns out that the two builds were very similar, but CMP did
not set build in default_makeflags, nor did it force non-parallel
make.

We must set build to avoid attempting to run on builder when
cross-building. It seems sane to also force non-parallel build, as
I don't trust the quality of EEMBC deliverables.

Change-Id: Ifb0c0d31984997287504d50a5d8990d1dc6e46ba