Pass COMPILER_FLAGS through default_makeflags
benchmark.sh puts COMPILER_FLAGS in the environment and expects
GNUmakefile to pick it up and pass it through.

This does not work for Coremark-Pro because COMPILER_FLAGS is
defined in its own makefiles. File-defined variables trump
environment-defined variables unless -e is set (and setting -e
is a recipe for chaos).

However, CLI-defined variables _do_ override file-defined variables.
So we transfer COMPILER_FLAGS from the environment to the CLI for
CLI for Coremark-Pro.

To maintain consistent behaviour for all benchmarks, we do the same
for the SPECs.

The right fix is to make benchmark.sh pass COMPILER_FLAGS through
makeflags rather than through the environment. This does not work
with current abe due to problems with the way it processes
command-line options.

Change-Id: Id4b42c2b6d8eefd7e8b94c8a3f7819141ca289b7