Generate YAML direct, not via template
Completely rework job generation to produce YAML on the fly, not
variables to be slotted into a job definition. More flexible, and
in particular makes it easy to have multiple targets for a single
host.

As far as possible, make the same ordering and whitespace choices
as old Benchmark.sh, for each of testing. May change those
choices in a later commit.

Change-Id: Ic51031483d5bd8310c28c160f907292a24ea3e63