Send most output to /dev/null
Not dumping on the console will speeds things up - at a minimum
it makes CPU2006 builds about 75% faster. We still capture the
output and deliver in results bundles, but it is no longer
possible to monitor build progress in web ui.

I believe LAVA uses the console for its own purposes, so don't
just kill all output in the .yaml files.

Don't have evidence that it helps to suppress on the target side,
so have left most output there. This allows users to see progress
in the web interface. But it could do with investigation, as
'seeing progress' is not worth affecting benchmark results and may
not be worth a cost in run time.

Did remove the 'set -x' from the gather scripts, which does affect
target size. This seems mostly noise once scripts are out of
development, so good to remove.

Change-Id: I9f27c904c3702191df5045ba69a1f6a06a3785a6