HACK: Give Jenkins artifacts time to get published
Until such time as we are able to dispatch the LAVA job _after_
the artifacts get published, we have a good chance of trying to
access them before they get published. Look out for Jenkins URLs
and give them up to 40 minutes to show up.

Will revert this shortly if I can find a better way, but the answer
may be to wait for an upgrade in Jenkins Job Builder that will
allow us to arrange our Jenkins jobs in a race-free manner. See
the huge comment in whichever job has ended up triggering regular
benchmark runs.

Change-Id: Ia445f5a2fe62b38a1028e7e49ab614b01e6b0c58