Call contrib/gcc_update just before building, not after checkout.
This is required to make sure the dependencies are up-to-date
before starting the build, whether stage1 or stage2.

Without this patch, gcc_update is called after cloning gcc,
but before checking out the requested branch/version.
The checkout may invalidate the result of gcc_update,
which has to been run before starting the build.

Change-Id: I6b6eff6592da3b6be7ae640da16a18ec7c08569a